--- grid_ref_dialog.py
@@ -1,41 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-/***************************************************************************
- GridRefDialog
-                                 A QGIS plugin
- This plugin takes the coords of the map cnavas and translates to an Ordnance Survey Grid Reference e.g. SX4855
-                             -------------------
-        begin                : 2014-08-21
-        git sha              : $Format:%H$
-        copyright            : (C) 2014 by Matt Travis
-        email                : matt.travis1@gmail.com
- ***************************************************************************/
-
-/***************************************************************************
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *                                                                         *
- ***************************************************************************/
-"""
-
-import os
-
-from PyQt4 import QtGui, uic
-
-FORM_CLASS, _ = uic.loadUiType(os.path.join(
-    os.path.dirname(__file__), 'grid_ref_dialog_base.ui'))
-
-
-class GridRefDialog(QtGui.QDialog, FORM_CLASS):
-    def __init__(self, parent=None):
-        """Constructor."""
-        super(GridRefDialog, self).__init__(parent)
-        # Set up the user interface from Designer.
-        # After setupUI you can access any designer object by doing
-        # self.<objectname>, and you can use autoconnect slots - see
-        # http://qt-project.org/doc/qt-4.8/designer-using-a-ui-file.html
-        # #widgets-and-dialogs-with-auto-connect
-        self.setupUi(self)
