--- grid_ref_dialog_base.ui
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ui version="4.0">
- <class>OSGBGridRef</class>
- <widget class="QDialog" name="OSGBGridRef">
-  <property name="geometry">
-   <rect>
-    <x>0</x>
-    <y>0</y>
-    <width>184</width>
-    <height>69</height>
-   </rect>
-  </property>
-  <property name="windowTitle">
-   <string>OSGB Grid Ref</string>
-  </property>
-  <widget class="QPushButton" name="pushButton">
-   <property name="geometry">
-    <rect>
-     <x>50</x>
-     <y>20</y>
-     <width>75</width>
-     <height>23</height>
-    </rect>
-   </property>
-   <property name="text">
-    <string>Get OS Ref</string>
-   </property>
-  </widget>
- </widget>
- <resources/>
- <connections>
-  <connection>
-   <sender>pushButton</sender>
-   <signal>pressed()</signal>
-   <receiver>OSGBGridRef</receiver>
-   <slot>accept()</slot>
-   <hints>
-    <hint type="sourcelabel">
-     <x>107</x>
-     <y>20</y>
-    </hint>
-    <hint type="destinationlabel">
-     <x>102</x>
-     <y>52</y>
-    </hint>
-   </hints>
-  </connection>
- </connections>
-</ui>
